fix installation of desktop files

+++ plasma-firstsetup.sh
@@ -80,17 +80,22 @@ fi
 
 #(tpg) add special icons on DESKTOP
 USER_DESKTOP=`xdg-user-dir DESKTOP`
-if [ ! -e $USER_DESKTOP/om-welcome.desktop ]; then
+if [ ! -e $USER_DESKTOP/om-welcome.desktop -a -e /usr/share/applications/om-welcome.desktop ]; then
     cp -f /usr/share/applications/om-welcome.desktop $USER_DESKTOP 2> /dev/null
 fi
 
-if [ ! -e $USER_DESKTOP/join.desktop ]; then
+if [ ! -e $USER_DESKTOP/join.desktop -a -e /usr/share/applications/join.desktop ]; then
     cp -f /usr/share/applications/join.desktop $USER_DESKTOP 2> /dev/null
 fi
 
-if [ ! -e $USER_DESKTOP/donate.desktop ]; then
+if [ ! -e $USER_DESKTOP/donate.desktop -a -e /usr/share/applications/donate.desktop ]; then
     cp -f /usr/share/applications/donate.desktop $USER_DESKTOP 2> /dev/null
 fi
+
+if [ ! -e $USER_DESKTOP/calamares.desktop -a -e /usr/share/applications/calamares.desktop ]; then
+    cp -f /usr/share/applications/calamares.desktop $USER_DESKTOP 2> /dev/null
+fi
+
 fi
 
 # GTK settings
